Some adding

+++ ft_ptintf/libft/libft.h
@@ -90,7 +90,7 @@ void			ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list			*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 
 void			ft_lstpush(t_list **alst, t_list *new);
-double			ft_pow(double num, long long power);
+long double		ft_pow(long double num, long long power);
 int				ft_isupper(int c);
 int				ft_islower(int c);
 int				ft_fib(int n);
